tietokannan alustus kunnossa
