--- viiteri/.env
@@ -1,1 +1,0 @@
-DATABASE_URL="postgresql+psycopg2://user"
