+++ viiteri/utils/.env
@@ -1,0 +1,1 @@
+DATABASE_URL=postgresql+psycopg2://
