+++ viiteri/utils/db.py
@@ -1,11 +1,14 @@
 from viiteri.app import app
 from flask_sqlalchemy import SQLAlchemy
+from dotenv import load_dotenv
 import os
 
+load_dotenv()
 
 app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv("DATABASE_URL")
 db = SQLAlchemy(app)
 
+app.app_context().push()
 
 def get_database_connection():
     return db
